Moved getLastPoint and setLastPoint to poinTRepository
