+++ backend/src/data_access_layer/point_repository.js
@@ -1,66 +1,81 @@
-const admin = require("firebase-admin")
+const fs = require('fs')
+const path = require('path')
 
-module.exports = function ({ db }) {
-    return {
-        /**
-           * @param {Map<String, Any>[]} callback
-           */
-        getAllPathPoints: function (callback) {
-            const docRef = db.collection('maps').doc("mapTest")
-            docRef.get().then((docSnap) => {
-                if (docSnap.empty) {
-                    callback('NoExistingMap', null)
-                } else {
-                    callback(null, docSnap.data().PathPoints)
-                }
-            }).catch((e) => {
-                console.log(e)
-                callback('InternalError', null)
-            })
-        },
+module.exports = function ({ db, admin }) {
+    function getAllPathPoints (callback) {
+        const docRef = db.collection('maps').doc("mapTest")
+        docRef.get().then((docSnap) => {
+            if (docSnap.empty) {
+                callback('NoExistingMap', null)
+            } else {
+                callback(null, docSnap.data().PathPoints)
+            }
+        }).catch((e) => {
+            console.log(e)
+            callback('InternalError', null)
+        })
+    }
 
-        /**
-           * @param {Map<Object, number>} coordinates
-           * @param {String} mapId
-           * @param {void} callback
-           */
-        getPointByCoordinate: function (mapId, coordinates, callback) {
-            const docRef = db.collection('maps').doc(mapId)
-            docRef.get().then((map) => {
-                const data = map.data()
-                if (data) {
-                    for (const point of data.points) {
-                        if (point.coordinates) {
-                            if (point.coordinates.x === coordinates.x && point.coordinates.y === coordinates.y) {
-                                callback(null, point)
-                                return
-                            }
+    function getPointByCoordinate (mapId, coordinates, callback) {
+        const docRef = db.collection('maps').doc(mapId)
+        docRef.get().then((map) => {
+            const data = map.data()
+            if (data) {
+                for (const point of data.points) {
+                    if (point.coordinates) {
+                        if (point.coordinates.x === coordinates.x && point.coordinates.y === coordinates.y) {
+                            callback(null, point)
+                            return
                         }
                     }
                 }
-                callback('PointNotFound', null)
-            }).catch(() => {
-                callback('DatabaseError', null)
+            }
+            callback('PointNotFound', null)
+        }).catch(() => {
+            callback('DatabaseError', null)
+        })
+    }
+
+    function addPoint (coordinates, callback) {
+        const docRef = db.collection('maps').doc('mapTest')
+        docRef.update({
+
+            points: admin.firestore.FieldValue.arrayUnion({
+                coordinates: coordinates
             })
-        },
+        }).then(() => {
+            callback(null, 200)
+        }).catch((error) => {
+            callback(error)
+        })
+    }
 
-        /**
-           * @param {Map<Number>} coordinates
-           * @param {Map<String, Any>} callback
-          */
-
-        addPoint: function (coordinates, callback) {
-            const docRef = admin.firestore().collection('maps').doc('mapTest')
-            docRef.update({
-
-                points: admin.firestore.FieldValue.arrayUnion({
-                    coordinates: coordinates
-                })
-            }).then(() => {
-                callback(null, 200)
-            }).catch((error) => {
-                callback(error)
-            })
+    function getLastPoint () {
+        const filePath = path.join(__dirname, 'last_point.json')
+        if (fs.existsSync(filePath)) {
+            const fileData = fs.readFileSync(filePath)
+            const dataObject = JSON.parse(fileData)
+            return dataObject
         }
+        return null
+    }
+
+    function setLastPoint (point) {
+        try {
+            const filePath = path.join(__dirname, 'last_point.json')
+            fs.writeFileSync(filePath, JSON.stringify(point))
+            return true
+        } catch (e) {
+            console.error(e)
+            return false
+        }
+    }
+
+    return {
+        getAllPathPoints,
+        getPointByCoordinate,
+        addPoint,
+        getLastPoint,
+        setLastPoint
     }
 }
